Update version to 1.5.0 and fix simultaneous display of process editor and toolbox
Changes:
- Update version from 1.0.0 to 1.5.0 in AssemblyInfo.cs
- Update ProductVersion and add ProductUpdateTime (2026-09-10) in Configuration.cs
- Display version and update time in debug title bar: [Debug: 2026-09-10]
- Fix ShowToolboxInVisionSidebar() to display toolbox in Document pane instead of sharing Frm_Job pane
- This allows process editor (DockRight) and toolbox (Document) to display side-by-side

+++ VisionAndMotion/3 FormLib/Frm_LayoutManage.cs
@@ -113,7 +113,7 @@ namespace VMPro
             {
                 string layoutName = dataGridView1.SelectedRows[0].Cells[1].Value.ToString() + ".config";
                 string layoutPath = "Config\\Resources\\Layout\\" + layoutName;
-                if (layoutName != "经典布局1.config" && layoutName != "经典布局2.config" && layoutName != "上次退出时布局.config")
+                if (layoutName != "经典布局1.config" && layoutName != "经典布局2.config" && layoutName != "左中右布局.config" && layoutName != "上次退出时布局.config")
                 {
                     if (File.Exists(Application.StartupPath + "\\" + layoutPath))
                         File.Delete(Application.StartupPath + "\\" + layoutPath);
@@ -193,7 +193,7 @@ namespace VMPro
             {
                 string layoutName = dataGridView1.SelectedRows[0].Cells[1].Value.ToString() + ".config";
                 string layoutPath = "Config\\Resources\\Layout\\" + layoutName;
-                if (layoutName != "经典布局1.config" && layoutName != "经典布局2.config" && layoutName != "上次退出时布局.config")
+                if (layoutName != "经典布局1.config" && layoutName != "经典布局2.config" && layoutName != "左中右布局.config" && layoutName != "上次退出时布局.config")
                 {
                     if (File.Exists(Application.StartupPath + "\\" + layoutPath))
                         File.Delete(Application.StartupPath + "\\" + layoutPath);
